ai_travel_agent/pyproject.toml
[project]
name = "ai_travel_agent"
version = "0.1.0"
description = "CrewAI-based multi-agent travel planner"
authors = [{ name = "Your Name", email = "your@email.com" }]
dependencies = [
  "absl-py==2.3.1",
"aiohappyeyeballs==2.6.1",
"aiohttp==3.13.0",
"aiosignal==1.4.0",
"annotated-types==0.7.0",
"anyio==4.11.0",
"appdirs==1.4.4",
"asttokens==3.0.0",
"astunparse==1.6.3",
"attrs==25.4.0",
"backoff==2.2.1",
"bcrypt==5.0.0",
"blinker==1.9.0",
"build==1.3.0",
"cachetools==6.2.0",
"certifi==2025.10.5",
"cffi==2.0.0",
"charset-normalizer==3.4.3",
"chromadb==1.1.1",
"click==8.3.0",
"coloredlogs==15.0.1",
"crewai==0.203.0",
"cryptography==46.0.2",
"decorator==5.2.1",
"diskcache==5.6.3",
"distro==1.9.0",
"docstring-parser==0.17.0",
"durationpy==0.10",
"et-xmlfile==2.0.0",
"executing==2.2.1",
"filelock==3.20.0",
"flatbuffers==25.9.23",
"frozenlist==1.8.0",
"fsspec==2025.9.0",
"gast==0.6.0",
"google-auth==2.41.1",
"google-pasta==0.2.0",
"googleapis-common-protos==1.70.0",
"grpcio==1.75.1",
"h11==0.16.0",
"h5py==3.14.0",
"hf-xet==1.1.10",
"httpcore==1.0.9",
"httptools==0.6.4",
"httpx==0.28.1",
"huggingface-hub==0.35.3",
"humanfriendly==10.0",
"idna==3.10",
"importlib-metadata==8.7.0",
"importlib-resources==6.5.2",
"instructor==1.11.3",
"ipython==9.6.0",
"ipython-pygments-lexers==1.1.1",
"jedi==0.19.2",
"jinja2==3.1.6",
"jiter==0.10.0",
"json-repair==0.25.2",
"json5==0.12.1",
"jsonpickle==4.1.1",
"jsonref==1.1.0",
"jsonschema==4.25.1",
"jsonschema-specifications==2025.9.1",
"keras==3.11.3",
"kubernetes==34.1.0",
"libclang==18.1.1",
"litellm==1.74.9",
"markdown==3.9",
"markdown-it-py==4.0.0",
"markupsafe==3.0.3",
"matplotlib-inline==0.1.7",
"mdurl==0.1.2",
"ml-dtypes==0.3.2",
"mmh3==5.2.0",
"mpmath==1.3.0",
"multidict==6.7.0",
"namex==0.1.0",
"networkx==3.5",
"numpy==1.26.4",
"oauthlib==3.3.1",
"onnxruntime==1.23.1",
"openai==1.109.1",
"openpyxl==3.1.5",
"opentelemetry-api",
"opentelemetry-exporter-otlp-proto-common",
"opentelemetry-exporter-otlp-proto-grpc",
"opentelemetry-exporter-otlp-proto-http",
"opentelemetry-proto",
"opentelemetry-sdk==1.37.0",
"opentelemetry-semantic-conventions==0.58b0",
"opt-einsum==3.4.0",
"optree==0.17.0",
"orjson==3.11.3",
"overrides==7.7.0",
"packaging==25.0",
"parso==0.8.5",
"pdfminer-six==20250506",
"pdfplumber==0.11.7",
"pexpect==4.9.0",
"pillow==11.3.0",
"portalocker==2.7.0",
"posthog==5.4.0",
"prompt-toolkit==3.0.52",
"propcache==0.4.1",
"protobuf",
"ptyprocess==0.7.0",
"pure-eval==0.2.3",
"pyasn1==0.6.1",
"pyasn1-modules==0.4.2",
"pybase64==1.4.2",
"pycparser==2.23",
"pydantic==2.12.0",
"pydantic-core==2.41.1",
"pydantic-settings==2.11.0",
"pygments==2.19.2",
"pyjwt==2.10.1",
"pypdfium2==4.30.0",
"pypika==0.48.9",
"pyproject-hooks==1.2.0",
"python-dateutil==2.9.0.post0",
"python-dotenv==1.1.1",
"pyvis==0.3.2",
"pyyaml==6.0.3",
"referencing==0.36.2",
"regex==2025.9.18",
"requests==2.32.5",
"requests-oauthlib==2.0.0",
"rich==14.2.0",
"rpds-py==0.27.1",
"rsa==4.9.1",
"setuptools==80.9.0",
"shellingham==1.5.4",
"six==1.17.0",
"sniffio==1.3.1",
"stack-data==0.6.3",
"sympy==1.14.0",
"tenacity==9.1.2",
"tensorboard==2.16.2",

"termcolor==3.1.0",
"tiktoken==0.12.0",
"tokenizers==0.22.1",
"tomli==2.3.0",
"tomli-w==1.2.0",
"tqdm==4.67.1",
"traitlets==5.14.3",
"typer==0.19.2",
"typing-extensions==4.15.0",
"typing-inspection==0.4.2",
"urllib3>=1.24.2,<2.4.0",
"uv==0.9.1",
"uvicorn==0.37.0",
"uvloop==0.21.0",
"watchfiles==1.1.0",
"wcwidth==0.2.14",
"websocket-client==1.9.0",
"websockets==15.0.1",
"werkzeug==3.1.3",
"wheel==0.45.1",
"wrapt==1.17.3",
"yarl==1.22.0",
"zipp==3.23.0"
]

[tool.setuptools]
packages = ["src"]  # If you're using setuptools

[tool.crewai]
entry = "src.ai_travel_agent:run_crew"
